show detail view when item is selected
